requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
--extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu121
torch>=2.1.0
llama-cpp-python
whisperx
pathvalidate
gradio
py7zr
pysubs2
